Added support for references
At the moment there should be a file `bib/bibliography.bib` containing
all bibliography entries. And we have to use Andreas Classen's bib2html
PHP script, which so far does not exist as a Git repository, hence not
as a submodule of this repository. Moreover, we have to remove the
superfluous <ol> from the output.